+++ src/fkooman/OAuth/Server/ClientData.php
@@ -246,8 +246,12 @@ class ClientData
         return $this->allowedScope;
     }
 
-    public function getSecret()
+    public function getSecret($getHash = false)
     {
+        if (null !== $this->secret && $getHash) {
+            return password_hash($this->secret, PASSWORD_DEFAULT);
+        }
+
         return $this->secret;
     }